### 2019-12-13 #1948 @ja-thomas
My colleague @christophM could do the review, he works as a researcher in interpretable machine learning.
