### 2019-12-17 #1948 @hiendn
I think that @frycast would be perfect as a reviewer for this paper. He is currently working on a Shapley values project of his own!
