### 2019-12-17 440cd85
FIxes missing dois in JOSS bib as requested by review (#168)
