+++ inst/joss_paper/paper.bib
@@ -31,7 +31,8 @@
   number={3},
   pages={647--665},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10115-013-0679-x},
 }
 
 @article{lundberg2018consistent,
@@ -49,7 +50,8 @@
   number={2},
   pages={271--293},
   year={1998},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/1467-9868.00125}
 }
 
 @Article{Shapley53,
@@ -76,7 +78,8 @@ journal = {{The Annals of Mathematical Statistics}},
 pages = {832--837},
 title = {{Remarks on Some Nonparametric Estimates of a Density Function}},
 volume = 27,
-year = 1956
+year = 1956,
+doi = {10.1214/aoms/1177728190}
 }
 
 
