+++ DESCRIPTION
@@ -9,8 +9,8 @@ Description: Complex machine learning models are often hard to interpret. Howeve
   described in Aas, Jullum and Løland (2019) <arXiv:1903.10464>, which accounts for any feature 
   dependence, and thereby produces more accurate estimates of the true Shapley values.
 Authors@R: c(
-    person("Nikolai", "Sellereite", email = "nikolaisellereite@gmail.com", role = c("cre", "aut"), comment = c(ORCID = 0000-0002-4671-0337)),
-    person("Martin", "Jullum", email = "Martin.Jullum@nr.no", role = "aut", comment = c(ORCID = 0000-0003-3908-5155)),  
+    person("Nikolai", "Sellereite", email = "nikolaisellereite@gmail.com", role = c("cre", "aut"), comment = c(ORCID = "0000-0002-4671-0337")),
+    person("Martin", "Jullum", email = "Martin.Jullum@nr.no", role = "aut", comment = c(ORCID = "0000-0003-3908-5155")),  
     person("Anders", "Løland", email = "Anders.Loland@nr.no", role = "ctb"), 
     person("Jens Christian", "Wahl", email = "Jens.Christian.Wahl@nr.no", role = "ctb"), 
     person("Camilla", "Lingjærde", role = "ctb"))
