### 2019-12-19 #1948 @martinju
DOI's should be fine now.
@terrytangyuan  do you need more help to find potential reviewers?
