### 2020-01-06 #1948 @martinju
> @ja-thomas There's potential conflict of interest according to @christophM so I am removing him for now. @ja-thomas Would you like to review this instead?

@ja-thomas ?
