### 2020-01-10 94dc646
Test fail running ctree
!is.null(dt[["id_combination"]]) is not TRUE, so some change in master for this must cause the error. Also, probably need to update the test objects.
