+++ man/prepare_data.Rd
@@ -14,15 +14,6 @@ prepare_data(x, ...)
 
 \method{prepare_data}{gaussian}(x, seed = 1, n_samples = 1000, index_features = NULL, ...)
 
-<<<<<<< HEAD
-\method{prepare_data}{copula}(x, x_test_gaussian = 1, seed = 1,
-  n_samples = 1000, index_features = NULL, ...)
-
-\method{prepare_data}{ctree}(x, seed = 1, n_samples = 1000,
-  index_features = NULL, mc_cores = 1,
-  mc_cores_simulateAllTrees = mc_cores,
-  mc_cores_sample_ctree = mc_cores, ...)
-=======
 \method{prepare_data}{copula}(
   x,
   x_test_gaussian = 1,
@@ -31,7 +22,17 @@ prepare_data(x, ...)
   index_features = NULL,
   ...
 )
->>>>>>> master
+
+\method{prepare_data}{ctree}(
+  x,
+  seed = 1,
+  n_samples = 1000,
+  index_features = NULL,
+  mc_cores = 1,
+  mc_cores_simulateAllTrees = mc_cores,
+  mc_cores_sample_ctree = mc_cores,
+  ...
+)
 }
 \arguments{
 \item{x}{Explainer object. See \code{\link{explain}} for more information.}
@@ -43,25 +44,18 @@ prepare_data(x, ...)
 \item{n_samples}{Positive integer. Indicating the maximum number of samples to use in the
 Monte Carlo integration for every conditional expectation.}
 
-<<<<<<< HEAD
-\item{index_features}{Positive integer vector. Only used internally.}
-
-\item{mc_cores}{Integer. Only for class \code{ctree}. The number of cores to use in paralellization of the
-tree building and tree prediction Defaults to 1. Uses parallel::mclapply which relies on forking, i.e. does not
-work on Windows systems.}
-
-\item{mc_cores_simulateAllTrees}{Integer. As for \code{mc_cores}, but specific for the tree building function
-#' Defaults to \code{mc_cores}.}
-
-\item{mc_cores_simulateAllTrees}{Integer. As for \code{mc_cores}, but specific for the tree building prediction function.
-Defaults to \code{mc_cores}.}
-=======
-\item{index_features}{Positive integer vector. Specifies the indices of combinations to apply to the present method.
-\code{NULL} means all combinations. Only used internally.}
+\item{index_features}{List. Default is NULL but if either various methods are being used or various mincriterion are used
+for different numbers of conditoned features, this will be a list with the features to pass.}
 
 \item{x_test_gaussian}{Matrix. Test data quantile-transformed to standard Gaussian variables. Only applicable if
 \code{approach = "empirical"}.}
->>>>>>> master
+
+\item{mc_cores}{Integer. Only for class \code{ctree} currently. The number of cores to use in paralellization of the
+tree building and tree prediction. Defaults to 1. Uses parallel::mclapply which relies on forking, i.e. does not
+work on Windows systems.}
+
+\item{mc_cores_simulateAllTrees}{Integer. Same as \code{mc_cores}, but specific for the tree building prediction function.
+Defaults to \code{mc_cores}.}
 }
 \description{
 Generate data used for predictions
