+++ NAMESPACE
@@ -38,7 +38,6 @@ export(prepare_data)
 export(rss_cpp)
 export(shapr)
 export(simulateAllTrees)
-export(weight_matrix)
 export(weight_matrix_cpp)
 importFrom(Rcpp,sourceCpp)
 importFrom(data.table,":=")
