### 2020-01-10 4781d88
script for NAV and description of functions of simulating true Shapley for categorical var
