+++ inst/devel_scripts/true_shapley/calculate_true_shapley.R
@@ -1,5 +1,22 @@
-
-## -------------------- some functions ----------------------
+#' Function to simulate Normal random variables with true mu and Sigma parameters.
+#'
+#' @description
+#'
+#' @param mu Numeric or vector indicating the true mean of the Normal or joint Normal random variables used to calculate
+#' the true Shapley values.
+#' @param Sigma Numeric (if mu is Numeric) or matrix with values used for covariance matrix of the random variables used to
+#' calculate the true Shapley values.
+#' @param beta Numeric or vector. These are the true coefficients of the response modelthat we are trying to explain with the
+#' Shapley values.
+#' @param N_shapley Numeric indicating how many Normal or joint Normal random variables to simulate. Default 10000.
+#' @param explainer explainer object from shapr package.
+#' @param cutoff vector of Numerics. This indicates where to cutoff the Normal random variables to make levels.
+#' @param response_mod function. The true response model that indicates how the features relate to the response.
+#' @details
+#'
+#' @return list
+#'
+#' @export
 
 sim_true_Normal <- function(mu, Sigma, beta, N_shapley = 10000, explainer, cutoff, response_mod){
 
@@ -54,6 +71,17 @@ sim_true_Normal <- function(mu, Sigma, beta, N_shapley = 10000, explainer, cutof
 
   return(list(joint_prob_dt, mn, prop))
 }
+
+#' Function to calculate marginal probabilities of the cutoff jointly Normal random variables
+#'
+#' @description
+#'
+#' @param joint_prob_dt list. Calculated using the \code{sim_true_Normal} function.
+#' @param explainer explainer object from shapr package.
+#'
+#' @return list
+#'
+#' @export
 
 marg_prob <- function(joint_prob_dt, explainer){
 
@@ -122,6 +150,18 @@ marg_prob <- function(joint_prob_dt, explainer){
   return(marg_list)
 }
 
+#' Function to calculate conditional probabilities of the cutoff jointly Normal random variables
+#'
+#' @description
+#'
+#' @param marg_list list. Contains the marginal probabilities calculated using the \code{marg_prob} function.
+#' @param joint_prob_dt list. Calculated using the \code{sim_true_Normal} function.
+#' @param explainer explainer object from shapr package.
+#'
+#' @return list
+#'
+#' @export
+
 cond_prob <- function(marg_list, joint_prob_dt, explainer){
 
   nms <- colnames(explainer$x_train)
@@ -176,8 +216,19 @@ cond_prob <- function(marg_list, joint_prob_dt, explainer){
   return(cond_list)
 }
 
+#' Function to calculate conditional expectations of the cutoff jointly Normal random variables
+#'
+#' @description
+#'
+#' @param x_test vector of test observations. Has the same dimension as the number of joint Normal random variables calculated in \code{sim_true_Normal} function.
+#' @param cond_list list. Calculated using the \code{cond_prob} function.
+#' @param explainer explainer object from shapr package.
+#' @param prediction_zero Numeric. Number to assigned to phi_0 in Shapley framework.
+#'
+#' @return list
+#'
+#' @export
 
-## function to calculate conditional expectation
 cond_expec <- function(x_test, cond_list, explainer, prediction_zero){ ## removed prediction_zero
 
   nms <- colnames(explainer$x_train)
@@ -219,6 +270,18 @@ cond_expec <- function(x_test, cond_list, explainer, prediction_zero){ ## remove
   return(cond_expec)
 }
 
[…]
