### 2020-01-14 38a2eac
Fixing bugs
nfeatures -> n_features, wcomb -> id_combination
