### 2020-01-14 c58a3b1
Safe update of explain test objects
