### 2020-01-14 e2f54e5
Fixes documentation warning
