### 2020-01-15 02673b1
All notes fixed locally
Had to move from .. prefix to the old ,with = FALSE for data.table to avoid a note
