+++ NAMESPACE
@@ -58,6 +58,7 @@ importFrom(data.table,year)
 importFrom(graphics,hist)
 importFrom(graphics,plot)
 importFrom(graphics,rect)
+importFrom(stats,as.formula)
 importFrom(stats,predict)
 importFrom(utils,head)
 importFrom(utils,tail)
