+++ tests/testthat/test-explanation.R
@@ -4,7 +4,7 @@ library(shapr)
 context("test-explanation.R")
 
 # For using same Random numer generator as CircelCI (R version 3.5.x)
-if (as.numeric(version$minor) >= 6.0) RNGkind(sample.kind = "Rounding")
+if (as.numeric(version$minor) >= 6.0 | as.numeric(version$major) > 3) RNGkind(sample.kind = "Rounding")
 
 test_that("Test functions in explanation.R", {
 
