+++ cran-comments.md
@@ -1,6 +1,8 @@
 ## Test environments
-* 
-* 
+* win-build (devel, release and oldrelease)
+* Windows Server 2008 R2 SP1, R-release, 32/64 bit, via R-hub
+* Windows Server 2008 R2 SP1, R-devel, 32/64 bit, via R-hub
+* Ubuntu Linux 16.04 LTS, R-release, GCC
 * 
 
 ## R CMD check results
