### 2020-01-15 c8fbf26
Ensure multicore test does not run on windows
