### 2020-01-15 6f41752
linting spaces and line breaks
styler::style_pkg(scope = "spaces") +
styler::style_pkg(scope = "line_breaks")
