+++ R/models.R
@@ -40,7 +40,6 @@ predict_model <- function(x, newdata) {
 #' @rdname predict_model
 #' @export
 predict_model.default <- function(x, newdata) {
-
   str_error <- paste(
     "It seems that you passed a non-valid model object.",
     "See more information about which models that are supported",
@@ -52,7 +51,6 @@ predict_model.default <- function(x, newdata) {
 #' @rdname predict_model
 #' @export
 predict_model.lm <- function(x, newdata) {
-
   if (!requireNamespace("stats", quietly = TRUE)) {
     stop("The stats package is required for predicting stats models")
   }
@@ -63,7 +61,6 @@ predict_model.lm <- function(x, newdata) {
 #' @rdname predict_model
 #' @export
 predict_model.glm <- function(x, newdata) {
-
   if (!requireNamespace("stats", quietly = TRUE)) {
     stop("The stats package is required for predicting stats models")
   }
@@ -78,7 +75,6 @@ predict_model.glm <- function(x, newdata) {
 #' @rdname predict_model
 #' @export
 predict_model.ranger <- function(x, newdata) {
-
   if (!requireNamespace("ranger", quietly = TRUE)) {
     stop("The ranger package is required for predicting ranger models")
   }
@@ -96,23 +92,20 @@ predict_model.ranger <- function(x, newdata) {
 #' @rdname predict_model
 #' @export
 predict_model.xgb.Booster <- function(x, newdata) {
-
   if (!requireNamespace("stats", quietly = TRUE)) {
     stop("The xgboost package is required for predicting xgboost models")
   }
-  if(model_type(x)=="cat_regression"){
+  if (model_type(x) == "cat_regression") {
     newdata_dummy <- predict(x$dummyfunc, newdata = newdata)
     predict(x, as.matrix(newdata_dummy))
   } else {
     predict(x, as.matrix(newdata))
   }
-
 }
 
 #' @rdname predict_model
 #' @export
 predict_model.gam <- function(x, newdata) {
-
   if (!requireNamespace("mgcv", quietly = TRUE)) {
     stop("The mgcv package is required for predicting gam models")
   }
@@ -126,7 +119,6 @@ predict_model.gam <- function(x, newdata) {
       predict(x, as.data.frame(newdata))
     )
   }
-
 }
 
 #' Define type of model
@@ -174,7 +166,6 @@ model_type.glm <- function(x) {
 #' @name model_type
 #' @export
 model_type.ranger <- function(x) {
-
   if (x$treetype == "Classification") {
     stop(
       paste0(
@@ -216,9 +207,8 @@ model_type.gam <- function(x) {
 #' @rdname model_type
 #' @export
 model_type.xgb.Booster <- function(x) {
-
   if (!is.null(x$params$objective) &&
-      (x$params$objective == "multi:softmax" | x$params$objective == "multi:softprob")
+    (x$params$objective == "multi:softmax" | x$params$objective == "multi:softprob")
   ) {
     stop(
       paste0(
@@ -243,10 +233,8 @@ model_type.xgb.Booster <- function(x) {
   ifelse(
     !is.null(x$params$objective) && x$params$objective == "binary:logistic",
     "classification",
-    ifelse(is.null(x$dummyfunc), "regression","cat_regression")
+    ifelse(is.null(x$dummyfunc), "regression", "cat_regression")
   )
-
-
 }
 
 #' Fetches feature labels from a given model object
@@ -264,7 +252,6 @@ features <- function(x, cnms, feature_labels = NULL) {
 
 #' @rdname features
 features.default <- function(x, cnms, feature_labels = NULL) {
-
[…]
