### 2020-01-15 5d3acf8
Extra manual linting fixed
