+++ R/sampling.R
@@ -222,14 +222,14 @@ sample_ctree <- function(tree,
 
     x_test_given <- x_test[, given_ind, drop = FALSE, with = FALSE]
 
-    xp <- x_test_given # data.table(matrix(x_test_given, nrow = 1, ncol = length(x_test_given)))  # this is changed by Martin
+    xp <- x_test_given # data.table(matrix(x_test_given, nrow = 1, ncol = length(x_test_given))) # change by MJ
     colnames(xp) <- paste0("V", given_ind) # this is important for where() below
 
     fit.nodes <- party::where(object = datact)
     ## I don't think you actually need this?
     # nodes <- unique(fit.nodes)
     # no.nodes <- length(nodes)
-    pred.nodes <- party::where(object = datact, newdata = xp) ## newdata must be a data.frame and have the same colnames as x
+    pred.nodes <- party::where(object = datact, newdata = xp) ## newdata must be data.frame +have the same colnames as x
 
     rowno <- 1:dim(x_train)[1]
 
@@ -242,7 +242,8 @@ sample_ctree <- function(tree,
 
     if (!sample) {
       if (length(rowno[fit.nodes == pred.nodes]) <= n_samples) {
-        depDT <- data.table::data.table(x_train[rowno[fit.nodes == pred.nodes], dependent_ind, drop = FALSE, with = FALSE])
+        depDT <- data.table::data.table(x_train[rowno[fit.nodes == pred.nodes], dependent_ind,
+                                                drop = FALSE, with = FALSE])
         givenDT <- data.table::data.table(x_test[1, given_ind, drop = FALSE, with = FALSE])
 
         ret <- cbind(depDT, givenDT)
@@ -299,7 +300,8 @@ sample_ctree <- function(tree,
 #' If \code{NULL}, the \code{mincriterion} is constant for every combination.
 #' This is depreciated and will be deleted soon.
 #'
-#' @param mincriterion Numeric value or vector equal to 1 - alpha where alpha is the nominal level of the conditional independence tests.
+#' @param mincriterion Numeric value or vector equal to 1 - alpha where alpha is the nominal level of the conditional
+#' independence tests.
 #' Can also be a vector equal to the length of the number of features indicating which mincriterion to use
 #' when conditioning on various numbers of features.
 #'
@@ -347,7 +349,8 @@ simulateAllTrees <- function(given_ind,
 
       colnames(df) <- c("Y", paste0("V", given_ind))
 
-      datact <- party::ctree(Y ~ ., data = df, controls = party::ctree_control(minbucket = minbucket, mincriterion = mincriterion))
+      datact <- party::ctree(Y ~ ., data = df, controls = party::ctree_control(minbucket = minbucket,
+                                                                               mincriterion = mincriterion))
     } else {
       x <- x_train[, given_ind, with = FALSE]
       y <- x_train[, dependent_ind, with = FALSE]
@@ -359,7 +362,8 @@ simulateAllTrees <- function(given_ind,
       ynam <- paste0("Y", 1:ncol(y))
       fmla <- as.formula(paste(paste(ynam, collapse = "+"), "~ ."))
 
-      datact <- party::ctree(fmla, data = df, controls = party::ctree_control(minbucket = minbucket, mincriterion = mincriterion))
+      datact <- party::ctree(fmla, data = df, controls = party::ctree_control(minbucket = minbucket,
+                                                                              mincriterion = mincriterion))
     }
   }
 
