### 2020-01-17 #2027 @whedon
[REVIEW]: shapr: An R-package for explaining machine learning models with dependence-aware Shapley values 
**Submitting author:** @martinju (<a href="http://orcid.org/0000-0003-3908-5155">Martin Jullum</a>)
**Repository:** <a href="https://github.com/NorskRegnesentral/shapr" target ="_blank">https://github.com/NorskRegnesentral/shapr</a>
**Version:** v0.1.1
