+++ DESCRIPTION
@@ -11,6 +11,7 @@ Description: Complex machine learning models are often hard to interpret. Howeve
 Authors@R: c(
     person("Nikolai", "Sellereite", email = "nikolaisellereite@gmail.com", role = c("cre", "aut"), comment = c(ORCID = "0000-0002-4671-0337")),
     person("Martin", "Jullum", email = "Martin.Jullum@nr.no", role = "aut", comment = c(ORCID = "0000-0003-3908-5155")),  
+    person("Annabelle", "Redelmeier", email = "Annabelle.Redelmeier@nr.no", role = "aut"),  
     person("Anders", "Løland", email = "Anders.Loland@nr.no", role = "ctb"), 
     person("Jens Christian", "Wahl", email = "Jens.Christian.Wahl@nr.no", role = "ctb"), 
     person("Camilla", "Lingjærde", role = "ctb"))
