### 2020-01-20 1230ea8
Update man after linting
