+++ man/gaussian_transform.Rd
@@ -18,7 +18,6 @@ Transforms a sample to standardized normal distribution
 \examples{
 y <- rnorm(50)
 x <- shapr:::gaussian_transform(y)
-
 }
 \author{
 Martin Jullum
