+++ man/simulateAllTrees.Rd
@@ -29,7 +29,8 @@ combination.
 If \code{NULL}, the \code{mincriterion} is constant for every combination.
 This is depreciated and will be deleted soon.}
 
-\item{mincriterion}{Numeric value or vector equal to 1 - alpha where alpha is the nominal level of the conditional independence tests.
+\item{mincriterion}{Numeric value or vector equal to 1 - alpha where alpha is the nominal level of the conditional
+independence tests.
 Can also be a vector equal to the length of the number of features indicating which mincriterion to use
 when conditioning on various numbers of features.}
 
@@ -45,7 +46,6 @@ Make all conditional inference trees
 }
 \examples{
 # TODO: Add simple example
-
 }
 \author{
 Annabelle Redelmeier
