+++ man/sample_combinations.Rd
@@ -34,7 +34,6 @@ nsamples <- 7
 joint_sampling <- FALSE
 cnms <- c("samp_train", "samp_test")
 x <- shapr:::sample_combinations(ntrain, ntest, nsamples, joint_sampling)
-
 }
 \author{
 Martin Jullum
