+++ man/inv_gaussian_transform.Rd
@@ -23,7 +23,6 @@ zx <- rnorm(50)
 n_z <- 30
 x <- shapr:::inv_gaussian_transform(zx, n_z)
 str(x)
-
 }
 \author{
 Martin Jullum
