### 2020-01-20 66d9795
generalizing ctree simulation study code
