+++ DESCRIPTION
@@ -22,7 +22,7 @@ Encoding: UTF-8
 LazyData: true
 ByteCompile: true
 Language: en-US
-RoxygenNote: 7.0.2
+RoxygenNote: 7.0.2  
 Depends: R (>= 3.5.0)
 Imports: 
     stats,
