+++ inst/joss_paper/paper.bib
@@ -31,7 +31,8 @@
   number={3},
   pages={647--665},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10115-013-0679-x},
 }
 
 @article{lundberg2018consistent,
@@ -49,7 +50,8 @@
   number={2},
   pages={271--293},
   year={1998},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/1467-9868.00125}
 }
 
 @Article{Shapley53,
@@ -76,13 +78,57 @@ journal = {{The Annals of Mathematical Statistics}},
 pages = {832--837},
 title = {{Remarks on Some Nonparametric Estimates of a Density Function}},
 volume = 27,
-year = 1956
+year = 1956,
+doi = {10.1214/aoms/1177728190}
 }
 
-@Manual{lime_api,
+
+  @Manual{limeRpackage,
     title = {lime: Local Interpretable Model-Agnostic Explanations},
     author = {Thomas Lin Pedersen and Michaël Benesty},
     year = {2019},
     note = {R package version 0.5.1},
     url = {https://CRAN.R-project.org/package=lime},
-}
+  }
+
+
+  @Manual{shapPythonpackage,
+    title = {shap: A unified approach to explain the output of any machine learning model},
+    author = {Scott Lundberg},
+    year = {2019},
+    note = {Python package version 0.32.1},
+    url = {https://pypi.org/project/shap/},
+  }
+
+  @Article{rcppRpackage,
+    title = {{Rcpp}: Seamless {R} and {C++} Integration},
+    author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
+    journal = {Journal of Statistical Software},
+    year = {2011},
+    volume = {40},
+    number = {8},
+    pages = {1--18},
+    url = {http://www.jstatsoft.org/v40/i08/},
+    doi = {10.18637/jss.v040.i08},
+  }
+
+  @Manual{rCore,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2019},
+    url = {https://www.R-project.org/},
+  }
+
+  @Manual{datatableRpackage,
+    title = {data.table: Extension of `data.frame`},
+    author = {Matt Dowle and Arun Srinivasan},
+    year = {2019},
+    note = {R package version 1.12.2},
+    url = {https://CRAN.R-project.org/package=data.table},
+  }
+
+
+
+
