### 2020-02-01 b6ad33c
typo "gassuian" -> "gaussian"
