### 2020-02-01 #2027 @frycast
This package looks great. The vignette is especially detailed and useful.

I have two comments/questions:
NorskRegnesentral/shapr/pull/172 - A small typo here.
NorskRegnesentral/shapr/pull/171 - You may need to add the full MIT license here.

Thank you @martinju for making this package!
