### 2020-02-01 NorskRegnesentral/shapr#173 @nikolase90
https://github.com/NorskRegnesentral/shapr/blob/991dfdb63a29d1c38d66c3352c00ca1a72764ac4/R/explanation.R#L20

This should be `gaussian` and not `gassuian`.
