### 2020-02-01 4ec87da
Fixes typo in R/explanation.R
