### 2020-02-01 NorskRegnesentral/shapr#175 @nikolase90
[review-frycast] LICENSE
@frycast The reason why the description file states the following
```
License: MIT + file LICENSE
```
is described in http://r-pkgs.had.co.nz/description.html#license. See the section with MIT license.

You could also see a similar discussion here: https://github.com/openjournals/joss-reviews/issues/1863.
