### 2020-02-01 #2027 @nikolase90
@frycast I've opened two issues based on your review.

- [ ] https://github.com/NorskRegnesentral/shapr/issues/173
- [ ] https://github.com/NorskRegnesentral/shapr/issues/175
