### 2020-02-01 NorskRegnesentral/shapr#176 @frycast
Vignette fails with uninformative error
When package gbm is missing, the [vignette](https://github.com/NorskRegnesentral/shapr/blob/master/vignettes/understanding_shapr.Rmd) fails to knit, and the error message doesn't make it very clear that gbm is the culprit.

I suggest  adding `library(gbm)` somewhere before line 399, so that a more informative error message is printed.
