### 2020-02-01 #2027 @frycast
>
>
> @frycast I've opened two issues based on your review.
>
>     * [ ]  [NorskRegnesentral/shapr#173](https://github.com/NorskRegnesentral/shapr/issues/173)
>
>     * [ ]  [NorskRegnesentral/shapr#175](https://github.com/NorskRegnesentral/shapr/issues/175)

Thanks @nikolase90 for resolving both of those issues, and thanks for writing this great package along with @martinju .

I've opened one more small issue:

NorskRegnesentral/shapr#176
