### 2020-02-01 NorskRegnesentral/shapr#176 @nikolase90
Thanks for catching that one!
