### 2020-02-01 9b7152b
Fixes issue when gbm is not installed
