### 2020-02-02 #2027 @nikolase90
Hi @frycast; https://github.com/NorskRegnesentral/shapr/pull/174 and https://github.com/NorskRegnesentral/shapr/pull/177 are now merged into master. Let us know if you have any further comments, and thanks for the nice comment about the package!
