### 2020-02-02 NorskRegnesentral/shapr#175 @nikolase90
@frycast Let us know if you have any further comments about this 👍 Otherwise we'll just close it and move forward. Thanks!
