### 2020-02-03 NorskRegnesentral/shapr#175 @martinju
Closed as @frycast approved the current license version, ref https://github.com/openjournals/joss-reviews/issues/2027#issuecomment-581192895
