### 2020-02-03 NorskRegnesentral/shapr#175 @martinju
Close #171 and this issue as @frycast approved the current license version, ref https://github.com/openjournals/joss-reviews/issues/2027#issuecomment-581192895
