### 2020-02-03 #2027 @martinju
> > Hi @frycast; [NorskRegnesentral/shapr#174](https://github.com/NorskRegnesentral/shapr/pull/174) and [NorskRegnesentral/shapr#177](https://github.com/NorskRegnesentral/shapr/pull/177) are now merged into master. Let us know if you have any further comments, and thanks for the nice comment about the package!
>
> Thank you @nikolase90 . I've ticked off my review checklist, and now we wait to hear from @expectopatronum .

Thanks for finishing the helpful review, @frycast .
I've now closed the license issue/PR.
