### 2020-02-03 76668f2
Made error-free version of groupSHAP
Next is testing
