### 2020-02-03 #2027 @expectopatronum
Hi @martinju,
there are some minor remarks regarding the references:
* Lundberg, S. (2019): I am not sure if this is correct/necessary. On their Github page they state which references should be used, this one doesn't seem to be there. (in case this actually should be there, I think SHAP should be capitalized)
* Maybe this also needs to be cited (very recent): https://raw.githubusercontent.com/slundberg/shap/master/docs/references/tree_explainer.bib
* Pedersen & Benesty: Title should be "lime: Local Interpretable Model-Agnostic Explanations"

Furthermore, capitalization of the titles and conference names is not consistent (some are capitalized, others are all lower case).
