### 2020-02-03 ffcb3b7
started testing grouping code
