### 2020-02-03 NorskRegnesentral/shapr#178 @martinju
[review-expectopatronum] References
Hi @expectopatronum
Just opening an issue to handle the reference comments from https://github.com/openjournals/joss-reviews/issues/2027#issuecomment-581468027

- [x] 1. Lundberg, S. (2019): I am not sure if this is correct/necessary. On their Github page they state which references should be used, this one doesn't seem to be there. (in case this actually should be there, I think SHAP should be capitalized)
- [x] 2. Maybe this also needs to be cited (very recent): https://raw.githubusercontent.com/slundberg/shap/master/docs/references/tree_explainer.bib
- [x] 3. Pedersen & Benesty: Title should be "lime: Local Interpretable Model-Agnostic Explanations"
- [x] 4. Furthermore, capitalization of the titles and conference names is not consistent (some are capitalized, others are all lower case).
