+++ inst/joss_paper/paper.bib
@@ -1,5 +1,5 @@
 @article{aas2019explaining,
-  title={Explaining individual predictions when features are dependent: More accurate approximations to Shapley values},
+  title={Explaining individual predictions when features are dependent: More accurate approximations to {S}hapley values},
   author={Aas, Kjersti and Jullum, Martin and L{\o}land, Anders},
   journal={arXiv preprint arXiv:1903.10464},
   year={2019}
@@ -56,8 +56,8 @@
 
 @Article{Shapley53,
 author={Lloyd S. Shapley},
-title={{A Value for N-Person Games}},
-journal={{Contributions to the Theory of Games}},
+title={A value for n-person games},
+journal={Contributions to the theory of games},
 volume=2,
 year = 1953,
 pages ={307--317}
@@ -84,21 +84,13 @@ doi = {10.1214/aoms/1177728190}
 
 
   @Manual{limeRpackage,
-    title = {lime: Local Interpretable Model-Agnostic Explanations},
+    title = {{lime}: Local Interpretable Model-Agnostic Explanations},
     author = {Thomas Lin Pedersen and Michaël Benesty},
     year = {2019},
     note = {R package version 0.5.1},
     url = {https://CRAN.R-project.org/package=lime},
   }
 
-
-  @Manual{shapPythonpackage,
-    title = {shap: A unified approach to explain the output of any machine learning model},
-    author = {Scott Lundberg},
-    year = {2019},
-    note = {Python package version 0.32.1},
-    url = {https://pypi.org/project/shap/},
-  }
 
   @Article{rcppRpackage,
     title = {{Rcpp}: Seamless {R} and {C++} Integration},
@@ -122,13 +114,30 @@ doi = {10.1214/aoms/1177728190}
   }
 
   @Manual{datatableRpackage,
-    title = {data.table: Extension of `data.frame`},
+    title = {{data.table}: Extension of `data.frame`},
     author = {Matt Dowle and Arun Srinivasan},
     year = {2019},
     note = {R package version 1.12.2},
     url = {https://CRAN.R-project.org/package=data.table},
   }
 
+@article{Lundberg2020,
+  author    = {Scott M. Lundberg and
+               Gabriel G. Erion and
+               Hugh Chen and
+               Alex DeGrave and
+               Jordan M. Prutkin and
+               Bala Nair and
+               Ronit Katz and
+               Jonathan Himmelfarb and
+               Nisha Bansal and
+               Su{-}In Lee},
+  title     = {From local explanations to global understanding with explainable {AI} for trees},
+  journal   = {Nature Machine Intelligence},
+  volume    = {2},
+  pages     = {56-67},
+  year      = {2020}
+}
 
 
 
