### 2020-02-03 NorskRegnesentral/shapr#178 @martinju
My attempt to fix the above issues can be found in PR #179. @expectopatronum please let me know if you find this satisfactory. The pdf can be found here: https://github.com/NorskRegnesentral/shapr/blob/references/inst/joss_paper/paper.pdf

Comments per issue:

1. I have removed this reference, and now rather mention the name of the Python package.
2. From my understanding the relevant part of this paper applies the TreeSHAP method already mentioned and cited (but now call it TreeExplainer instead). However, since this is a proper journal paper, I think I think it is a good idea to include it.
3. Fixed
4. I went through all citations and fixed uppercase/lowercase issues.
