### 2020-02-03 #2027 @martinju
Hi @expectopatronum

Thank you for raising these issues.

I've opened an issue for these at https://github.com/NorskRegnesentral/shapr/pull/178 (let's do any further discussion on there) and an associated PR at https://github.com/NorskRegnesentral/shapr/pull/179

The pdf for the PR-version of the paper https://github.com/NorskRegnesentral/shapr/blob/references/inst/joss_paper/paper.pdf

Please let me know if you find this satisfactory.
