+++ inst/joss_paper/paper.bib
@@ -16,7 +16,7 @@
 @article{kononenko2010efficient,
   title={An efficient explanation of individual classifications using game theory},
   author={{\v{S}}trumbelj, Erik and Kononenko, Igor},
-  journal={Journal of Machine Learning Research},
+  journal={Journal of machine learning research},
   volume={11},
   number={Jan},
   pages={1--18},
