+++ inst/joss_paper/paper.bib
@@ -95,7 +95,7 @@ doi = {10.1214/aoms/1177728190}
   @Article{rcppRpackage,
     title = {{Rcpp}: Seamless {R} and {C++} Integration},
     author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
-    journal = {Journal of Statistical Software},
+    journal = {Journal of statistical software},
     year = {2011},
     volume = {40},
     number = {8},
@@ -133,7 +133,7 @@ doi = {10.1214/aoms/1177728190}
                Nisha Bansal and
                Su{-}In Lee},
   title     = {From local explanations to global understanding with explainable {AI} for trees},
-  journal   = {Nature Machine Intelligence},
+  journal   = {Nature machine intelligence},
   volume    = {2},
   pages     = {56-67},
   year      = {2020}
