### 2020-02-03 NorskRegnesentral/shapr#178 @expectopatronum
1, 2, 3 look good!
4. there is still a mix of lower case and capitalization for conference/journal names
