### 2020-02-03 NorskRegnesentral/shapr#180 @expectopatronum
Question: Interpretation of output
I'm currently playing around with the example from the README and I am not sure if everything is as clear as it could be. Could you explain what `none` means in the output or point the user to the description?
```r
print(explanation$dt)
#>      none     lstat         rm       dis      indus
#> 1: 22.446 5.2632030 -1.2526613 0.2920444  4.5528644
#> 2: 22.446 0.1671903 -0.7088405 0.9689007  0.3786871
#> 3: 22.446 5.9888016  5.5450861 0.5660136 -1.4304350
#> 4: 22.446 8.2142203  0.7507569 0.1893368  1.8298305
#> 5: 22.446 0.5059890  5.6875106 0.8432240  2.2471152
#> 6: 22.446 1.9929674 -3.6001959 0.8601984  3.1510531
```

Or you could call it 'average prediction' and also show the difference, like [here](https://christophm.github.io/interpretable-ml-book/shapley.html)?
