### 2020-02-03 #2027 @expectopatronum
Regarding community guidelines: Could you add a hint where questions should be put? (most likely as a Github issue?)
