### 2020-02-03 #2027 @expectopatronum
Regarding your vignette (part 1 - to be continued tomorrow):
* Just a minor remark (since it is not an issue with your package but with devtools): If you want user to be able to use `vignette(shapr)`, they need to install the package using `devtools::install_github("NorskRegnesentral/shapr", build_vignettes = TRUE)`, because apparently [devtools does not build vignettes automatically](https://stackoverflow.com/a/33617870/1117932).
* Building the vignette failed because of missing requirment `caret`, you might need to add it to your requirements? Same will happen for `xgboost` (I had to install it in order to run your example from the README), and `gbm`.
* For opening the vignette the user first has to figure out it's name, naming it 'shapr' would make that easier. Or mention in the description that one has to call `vignette("understanding_shapr")`.
