### 2020-02-03 #2027 @nikolase90
Hi @expectopatronum!

We'll add a remark that if you want to actually want to build the vignette and README you'll have to slightly modify the installment procedure.

By doing
```
devtools::install_github("NorskRegnesentral/shapr", dependencies = TRUE)
```
you'll also install packages listed under `Suggets` in `DESCRIPTION ` (where `xgboost` and `caret` is listed). By doing this we avoid the users to download all the different packages.  The default value is `dependencies = NA`, which only installs the packages listed under `Imports`,  `Imports` and `LinkingTo`.
