### 2020-02-03 #2027 @terrytangyuan
@nikolase90 I just realized that there are many optional dependencies in DESCRIPTION. Are there any dependent packages that are worth being cited?
