### 2020-02-03 #2027 @nikolase90
@terrytangyuan Most of the packages listed under `Suggests` are not necessary to cite. Though we could add a citation for `RcppArmadillo` (since that is part of the core of our package). In addition to this we could also site the packages of the models we support natively.
