### 2020-02-04 #2027 @expectopatronum
Vignette (part 2):

* The citations look "weird", I tried to figure out whether I build the vignette wrong or it needs changing.
![image](https://user-images.githubusercontent.com/2265475/73718117-e4690b80-471b-11ea-9a2a-d66779a635e5.png)
If it looks ok for you, then ignore this comment :) (maybe you know how I can compile it correctly)

* Content:
> and that we want to explain the predictions for new sets of features.

shouldn't this be:

> and that we want to explain the predictions for new sets of examples. ?

* Formatting of the following formula (again, if I built the vignette incorrectly, never mind)
![image](https://user-images.githubusercontent.com/2265475/73718323-740eba00-471c-11ea-958b-250240cd06f1.png)

* Formatting again and it probably should say KernSHAP instead of KerSHAP

![image](https://user-images.githubusercontent.com/2265475/73718365-9bfe1d80-471c-11ea-881f-59a85dd37390.png)

Besides those remarks the vignette is really great and helpful!

After https://github.com/NorskRegnesentral/shapr/issues/178, https://github.com/NorskRegnesentral/shapr/issues/180 and https://github.com/NorskRegnesentral/shapr/issues/181 are resolved I am good to go!
