### 2020-02-04 NorskRegnesentral/shapr#178 @nikolase90
@martinju There is a mix of capital and lower cases where we cite R? Could you fix that?
