### 2020-02-04 87bea93
Enforced super strict convention on uppercase/lowercase
