+++ inst/joss_paper/paper.bib
@@ -1,7 +1,7 @@
 @article{aas2019explaining,
   title={Explaining individual predictions when features are dependent: More accurate approximations to {S}hapley values},
   author={Aas, Kjersti and Jullum, Martin and L{\o}land, Anders},
-  journal={arXiv preprint arXiv:1903.10464},
+  journal={Arxiv preprint arxiv:1903.10464},
   year={2019}
 }
 
@@ -38,7 +38,7 @@
 @article{lundberg2018consistent,
   title={Consistent individualized feature attribution for tree ensembles},
   author={Lundberg, Scott M and Erion, Gabriel G and Lee, Su-In},
-  journal={arXiv preprint arXiv:1802.03888},
+  journal={Arxiv preprint arxiv:1802.03888},
   year={2018}
 }
 
@@ -105,10 +105,9 @@ doi = {10.1214/aoms/1177728190}
   }
 
   @Manual{rCore,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {R: A language and environment for statistical computing},
     author = {{R Core Team}},
-    organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
+    organization = {R foundation for statistical computing},
     year = {2019},
     url = {https://www.R-project.org/},
   }
