### 2020-02-04 NorskRegnesentral/shapr#178 @martinju
I found no specific guidelines on how to format the BibTex entries for JOSS, but now enforced a strict convention of using uppercase for the first word in paper titles/journals and after colons + abbreviations like AI. For programming languages and packages I used the convention enforced by the authors (i.e. R, C++, Rcpp, data.table). Everything is all
lowercase, no matter the convention used by the journals themselves.

For citing R I had used the bibTeX entry provided by R itself, running citation():

  @Manual{,
    title = {R: A Language and Environment for Statistical Computing},
    author = {{R Core Team}},
    organization = {R Foundation for Statistical Computing},
    address = {Vienna, Austria},
    year = {2019},
    url = {https://www.R-project.org/},
  }

In the latest version I deleted the address field and force lowercased the organization (and title). R Core Team is the author, so that is left unchanged.

@expectopatronum I am sorry to ask you to look over the references once again here: https://github.com/NorskRegnesentral/shapr/blob/references/inst/joss_paper/paper.pdf
I sincerely hope this is satisfactory. If not, I kindly ask you to pinpoint the exact issues.
