### 2020-02-04 NorskRegnesentral/shapr#178 @expectopatronum
Looks perfect! Thanks!
