### 2020-02-04 #2027 @martinju
Thanks for the vignette comments @expectopatronum
We will improve the vignette installation instructions and look into the potential typo.
Btw, we also have a html version of the vignette at our pckdown site: https://norskregnesentral.github.io/shapr/articles/understanding_shapr.html

To actually build the vignette with `install_github` you also need the argument `build_vignette = T`

There have been some changes in devtools/remotes regarding defaults that are relevant if you want to build the vignette, but for versions newer than 2.1.0 and older than 2.0.0 (see https://github.com/r-lib/remotes/issues/353), you should install the package with

`devtools::install_github("NorskRegnesentral/shapr",dependencies = T,build_vignettes = T)`

if you want to install the package (and build the vignette at installation)
To display the vignette, then simply run
`vignette("understanding_shapr",package="shapr")`
Does this solve your formatting problem?
