### 2020-02-04 e3fbe60
Fixing references issues (#179)
