### 2020-02-04 NorskRegnesentral/shapr#178 @nikolase90
Fixed with #179 .
