### 2020-02-04 NorskRegnesentral/shapr#180 @martinju
Hi @expectopatronum

Thanks for the question. Yes, we should make this clearer. The first column is the part of the prediction value which is not "due to" any of the features, hence "none".
By design, the user can set this value. We prefer the mean response in the training data, while others use the mean prediction in the training data.

While it would change the interpretation a bit, it is also possible to use other reference levels, like the mean prediction in a subset of the training or the data to be explained data.

Other names like "reference level" could have been used, but I think "none" is fine when we provide short explanations of it in the help file of `shapr::explain`, the README and the vignette.

How does that sound?
