### 2020-02-04 dd1866b
Cites RcppArmadillo in paper
