+++ inst/joss_paper/paper.bib
@@ -138,5 +138,13 @@ doi = {10.1214/aoms/1177728190}
   year      = {2020}
 }
 
-
+@article{eddelbuettel2014rcpparmadillo,
+  title={RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
+  author={Eddelbuettel, Dirk and Sanderson, Conrad},
+  journal={Computational Statistics \& Data Analysis},
+  volume={71},
+  pages={1054--1063},
+  year={2014},
+  publisher={Elsevier}
+}
 
