+++ inst/joss_paper/paper.bib
@@ -141,7 +141,7 @@ doi = {10.1214/aoms/1177728190}
 @article{eddelbuettel2014rcpparmadillo,
   title={RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
   author={Eddelbuettel, Dirk and Sanderson, Conrad},
-  journal={Computational Statistics \& Data Analysis},
+  journal={Computational statistics \& data analysis},
   volume={71},
   pages={1054--1063},
   year={2014},
