### 2020-02-04 9c21b73
Update inst/joss_paper/paper.bib
Co-Authored-By: Martin Jullum <jullum@nr.no>
