+++ inst/joss_paper/paper.bib
@@ -139,7 +139,7 @@ doi = {10.1214/aoms/1177728190}
 }
 
 @article{eddelbuettel2014rcpparmadillo,
-  title={RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
+  title={RcppArmadillo: {A}ccelerating {R} with high-performance {C}++ linear algebra},
   author={Eddelbuettel, Dirk and Sanderson, Conrad},
   journal={Computational statistics \& data analysis},
   volume={71},
@@ -147,4 +147,3 @@ doi = {10.1214/aoms/1177728190}
   year={2014},
   publisher={Elsevier}
 }
-
