### 2020-02-04 cc396cc
Adds citations for packages
