### 2020-02-04 5901393
Cites RcppArmadillo in paper (#186)
