### 2020-02-04 b457775
Update inst/joss_paper/paper.Rmd
Co-Authored-By: Martin Jullum <jullum@nr.no>
