### 2020-02-04 #2027 @nikolase90
@terrytangyuan https://github.com/NorskRegnesentral/shapr/commit/5901393952249b3544736e1e3496d41b5cc35c55 and https://github.com/NorskRegnesentral/shapr/commit/13f1d460d3d00dcfb62742597875f8890f71b8eb adds citations for additional packages.
