+++ README.md
@@ -147,7 +147,10 @@ explanation <- explain(
   prediction_zero = p
 )
 
-# Printing the Shapley values for the test data
+# Printing the Shapley values for the test data.
+# Each row contains the Shapley values decomposition of each prediction.
+# The first column contains phi_0, the Shapley value not assigned to any features, defined as p above.
+# The remaining columns contain the Shapley values specific to the corresponding features.
 print(explanation$dt)
 #>      none     lstat         rm       dis      indus
 #> 1: 22.446 5.2632030 -1.2526613 0.2920444  4.5528644
