+++ README.Rmd
@@ -118,7 +118,10 @@ explanation <- explain(
   prediction_zero = p
 )
 
-# Printing the Shapley values for the test data
+# Printing the Shapley values for the test data. 
+# Each row contains the Shapley values decomposition of each prediction.
+# The first column contains phi_0, the Shapley value not assigned to any features, defined as p above. 
+# The remaining columns contain the Shapley values specific to the corresponding features.
 print(explanation$dt)
 
 # Finally we plot the resulting explanations
