+++ vignettes/understanding_shapr.Rmd
@@ -245,20 +245,22 @@ model <- xgboost(
 explainer <- shapr(x_train, model)
 
 # Specifying the phi_0, i.e. the expected prediction without any features
-pred_zero <- mean(y_train)
+p <- mean(y_train)
 
-# Computing the actual Shapley values with kernelSHAP accounting for feature 
-# dependence using the empirical (conditional) distribution approach with 
-# bandwidth parameter sigma = 0.1 (default)
+# Computing the actual Shapley values with kernelSHAP accounting for feature dependence using
+# the empirical (conditional) distribution approach with bandwidth parameter sigma = 0.1 (default)
 explanation <- explain(
   x_test,
   approach = "empirical",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
-# Printing the Shapley values for the test data
-explanation$dt
+# Printing the Shapley values for the test data. 
+# Each row contains the Shapley values decomposition of each prediction.
+# The first column contains phi_0, the Shapley value not assigned to any features, defined as p above. 
+# The remaining columns contain the Shapley values specific to the corresponding features.
+print(explanation$dt)
 
 # Plot the resulting explanations for observations 1 and 6
 plot(explanation, plot_phi0 = FALSE, index_x_test = c(1,6))
