### 2020-02-04 NorskRegnesentral/shapr#180 @nikolase90
@martinju I think it would be nice if we could add a section where we explain how you should interpret the `none`-column. It should be placed a the bottom of the section where we describe the output from `shapr::explain`. We do describe it for the features, but don't really say anything about the first column. Agree?
