### 2020-02-04 NorskRegnesentral/shapr#180 @martinju
Working on it. I'll push a PR in a few minutes.
