### 2020-02-04 NorskRegnesentral/shapr#180 @nikolase90
Ok, sweet.
