### 2020-02-04 1ca649f
pred_zero -> p as used elsewhere
