+++ vignettes/understanding_shapr.Rmd
@@ -274,7 +274,7 @@ explanation_gaussian <- explain(
   x_test,
   approach = "gaussian",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6
@@ -289,7 +289,7 @@ explanation_copula <- explain(
   x_test,
   approach = "copula",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6, excluding 
@@ -363,7 +363,7 @@ explanation_combined <- explain(
   x_test,
   approach = c("empirical","copula","gaussian","gaussian"),
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6, excluding 
