+++ vignettes/understanding_shapr.Rmd
@@ -229,8 +229,8 @@ data("Boston", package = "MASS")
 x_var <- c("lstat", "rm", "dis", "indus")
 y_var <- "medv"
 
-x_train <- as.matrix(Boston[-(1:6), x_var])
-y_train <- Boston[-(1:6), y_var]
+x_train <- as.matrix(Boston[-1:-6, x_var])
+y_train <- Boston[-1:-6, y_var]
 x_test <- as.matrix(Boston[1:6, x_var])
 
 # Fitting a basic xgboost model to the training data
