+++ README.Rmd
@@ -88,8 +88,8 @@ data("Boston", package = "MASS")
 x_var <- c("lstat", "rm", "dis", "indus")
 y_var <- "medv"
 
-x_train <- as.matrix(Boston[-(1:6), x_var])
-y_train <- Boston[-(1:6), y_var]
+x_train <- as.matrix(Boston[-1:-6, x_var])
+y_train <- Boston[-1:-6, y_var]
 x_test <- as.matrix(Boston[1:6, x_var])
 
 # Looking at the dependence between the features
@@ -118,9 +118,9 @@ explanation <- explain(
   prediction_zero = p
 )
 
-# Printing the Shapley values for the test data. 
+# Printing the Shapley values for the test data.
 # Each row contains the Shapley values decomposition of each prediction.
-# The first column contains phi_0, the Shapley value not assigned to any features, defined as p above. 
+# The first column contains phi_0, the Shapley value not assigned to any features, defined as p above.
 # The remaining columns contain the Shapley values specific to the corresponding features.
 print(explanation$dt)
 
