### 2020-02-04 8cb58f8
Updates README with installation instructions (#189)
