### 2020-02-04 00cfef5
Update R/explanation.R
Co-Authored-By: Nikolai Sellereite <n_sellereite@hotmail.com>
