### 2020-02-04 5e8371c
simplify readme comments
