### 2020-02-04 589a9e5
Update R/explanation.R
Co-Authored-By: Nikolai Sellereite <n_sellereite@hotmail.com>
