+++ inst/devel_scripts/true_shapley/simulation_true_shapley_withdatatable_MJ.R
@@ -1,0 +1,118 @@
+library(shapr)
+library(data.table)
+library(MASS)
+library(lqmm) ## to check if Sigma is positive definite
+library(rapportools) # for testing booleans
+
+# source("/nr/project/stat/BigInsight/Projects/Fraud/Subprojects/NAV/Annabelle/shapr/inst/devel_scripts/true_shapley/calculate_true_shapley.R")
+
+source("inst/devel_scripts/true_shapley/calculate_true_shapley_withdatatable.R")
+
+tod_date <- '17_01_20'
+
+response_mod = function(tbl, beta, mod_matrix){
+  # beta <- c(1.0, 1,  0.0,  0,  0,  0,  0,  0,  0, 0)
+  nms <- colnames(mod_matrix)[-1]
+
+  for(i in 1:length(nms)){
+    assign(nms[i], tbl[, nms[i], with = FALSE])
+  }
+
+  epsilon <- tbl[, "epsilon"]
+
+  return(beta[1] + beta[2] * ((1 - feat12) * (1 - feat13)) +  beta[3] * feat12 + beta[4] * feat13 +
+           beta[5] * ((1 - feat22) * (1 - feat23)) + beta[6] * feat22 + beta[7] * feat23 +
+           beta[8] * ((1 - feat32) * (1 - feat33)) + beta[9] * feat32 + beta[10] * feat33 + epsilon)
+}
+
+
+parameters_list <- list(Sigma_diag = 1,
+                        corr = 0,
+                        mu = c(0, 0, 0),
+                        beta = c(1, -1, 0, 1, 1, 1, 0.5, 0.5, 1, -1),
+                        N_shapley = 10000,
+                        N_training = 1000,
+                        N_testing = 1000,
+                        cutoff = c(-200, 0, 1, 200),
+                        noise = FALSE,
+                        response_mod = response_mod,
+                        fit_mod = "regression",
+                        methods = c("empirical"), # "gaussian", "ctree"
+                        name = 'corr0')
+
+# ll <- simulate_data(parameters_list)
+# head(ll$true_shapley)
+# head(ll$true_linear)
+# MAE(ll$true_shapley, ll$true_linear) # 0.00095050 for 10000 obs and 1000 for training and testing / 0.0009436
+# MAE(ll$true_shapley, ll$methods[['gaussian']]$dt_sum) # 0.01729 # 2.68 minutes
+# MAE(ll$true_shapley, ll$methods[['empirical']]$dt_sum) # 0.0179 # 1.06 minutes
+# MAE(ll$true_shapley, ll$methods[['ctree']]$dt) # 0.016423 #
+# MAE(ll$true_shapley, ll$methods[['ctree_onehot']]$dt_sum) # 0.013 5.46 minutes
+# MAE(ll$true_shapley, ll$methods[['empirical_ind']]$dt_sum) # 0.01876 # 2.10 minutes
+
+
+##
+
+
+
+tm <- Sys.time()
+all_methods <- list()
+for(i in 1:length(parameters_list)){ # length(parameters_list)
+  all_methods[[i]] <- simulate_data(parameters_list[[i]])
+  nm = paste(tod_date, '_results_', i,".rds", sep = "")
+  saveRDS(all_methods, file = paste("/nr/project/stat/BigInsight/Projects/Fraud/Subprojects/NAV/Annabelle/results/MAE/", nm, sep = ""))
+}
+
+mydata <- readRDS("/nr/project/stat/BigInsight/Projects/Fraud/Subprojects/NAV/Annabelle/results/MAE/17_01_20_results_7.rds")
+
+
+MAE_truth <- NULL
+MAE_methods <- NULL
[…]
