### 2020-02-04 75d7335
Fixed response_mod thing
Do we need the one-hot encoding for anything? Other similar cases like response_mod. linear_kshap?
