+++ README.Rmd
@@ -103,9 +103,10 @@ data("Boston", package = "MASS")
 x_var <- c("lstat", "rm", "dis", "indus")
 y_var <- "medv"
 
-x_train <- as.matrix(Boston[-1:-6, x_var])
-y_train <- Boston[-1:-6, y_var]
-x_test <- as.matrix(Boston[1:6, x_var])
+ind_x_test <- 1:6
+x_train <- as.matrix(Boston[-ind_x_test, x_var])
+y_train <- Boston[-ind_x_test, y_var]
+x_test <- as.matrix(Boston[ind_x_test, x_var])
 
 # Looking at the dependence between the features
 cor(x_train)
