### 2020-02-04 624344c
Document the none-column (#190)
