+++ inst/scripts/readme_example.R
@@ -1,3 +1,4 @@
+library(xgboost)
 library(shapr)
 
 data("Boston", package = "MASS")
@@ -5,15 +6,15 @@ data("Boston", package = "MASS")
 x_var <- c("lstat", "rm", "dis", "indus")
 y_var <- "medv"
 
-x_train <- as.matrix(tail(Boston[, x_var], -6))
-y_train <- tail(Boston[, y_var], -6)
-x_test <- as.matrix(head(Boston[, x_var], 6))
+x_train <- as.matrix(Boston[-1:-6, x_var])
+y_train <- Boston[-1:-6, y_var]
+x_test <- as.matrix(Boston[1:6, x_var])
 
-# Just looking at the dependence between the features
+# Looking at the dependence between the features
 cor(x_train)
 
 # Fitting a basic xgboost model to the training data
-model <- xgboost::xgboost(
+model <- xgboost(
   data = x_train,
   label = y_train,
   nround = 20,
@@ -23,15 +24,21 @@ model <- xgboost::xgboost(
 # Prepare the data for explanation
 explainer <- shapr(x_train, model)
 
-# Spedifying the phi_0, i.e. the expected prediction without any features
-p0 <- mean(y_train)
+# Specifying the phi_0, i.e. the expected prediction without any features
+p <- mean(y_train)
 
 # Computing the actual Shapley values with kernelSHAP accounting for feature dependence using
 # the empirical (conditional) distribution approach with bandwidth parameter sigma = 0.1 (default)
-explanation <- explain(x_test, explainer, approach = "empirical", prediction_zero = p0)
+explanation <- explain(
+  x_test,
+  approach = "empirical",
+  explainer = explainer,
+  prediction_zero = p
+)
 
-# Printing the Shapley values for the test data
-explanation$dt
+# Printing the Shapley values for the test data.
+# For more information about the interpretation of the values in the table, see ?shapr::explain.
+print(explanation$dt)
 
 # Finally we plot the resulting explanations
 plot(explanation)
