+++ vignettes/understanding_shapr.Rmd
@@ -229,8 +229,8 @@ data("Boston", package = "MASS")
 x_var <- c("lstat", "rm", "dis", "indus")
 y_var <- "medv"
 
-x_train <- as.matrix(Boston[-(1:6), x_var])
-y_train <- Boston[-(1:6), y_var]
+x_train <- as.matrix(Boston[-1:-6, x_var])
+y_train <- Boston[-1:-6, y_var]
 x_test <- as.matrix(Boston[1:6, x_var])
 
 # Fitting a basic xgboost model to the training data
@@ -245,20 +245,20 @@ model <- xgboost(
 explainer <- shapr(x_train, model)
 
 # Specifying the phi_0, i.e. the expected prediction without any features
-pred_zero <- mean(y_train)
+p <- mean(y_train)
 
-# Computing the actual Shapley values with kernelSHAP accounting for feature 
-# dependence using the empirical (conditional) distribution approach with 
-# bandwidth parameter sigma = 0.1 (default)
+# Computing the actual Shapley values with kernelSHAP accounting for feature dependence using
+# the empirical (conditional) distribution approach with bandwidth parameter sigma = 0.1 (default)
 explanation <- explain(
   x_test,
   approach = "empirical",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
-# Printing the Shapley values for the test data
-explanation$dt
+# Printing the Shapley values for the test data. 
+# For more information about the interpretation of the values in the table, see ?shapr::explain.  
+print(explanation$dt)
 
 # Plot the resulting explanations for observations 1 and 6
 plot(explanation, plot_phi0 = FALSE, index_x_test = c(1,6))
@@ -272,7 +272,7 @@ explanation_gaussian <- explain(
   x_test,
   approach = "gaussian",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6
@@ -287,7 +287,7 @@ explanation_copula <- explain(
   x_test,
   approach = "copula",
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6, excluding 
@@ -361,7 +361,7 @@ explanation_combined <- explain(
   x_test,
   approach = c("empirical","copula","gaussian","gaussian"),
   explainer = explainer,
-  prediction_zero = pred_zero
+  prediction_zero = p
 )
 
 # Plot the resulting explanations for observations 1 and 6, excluding 
