### 2020-02-04 #2027 @martinju
@expectopatronum

We now consider NorskRegnesentral/shapr#178, NorskRegnesentral/shapr#180, and NorskRegnesentral/shapr#181 resolved in master.

Please let us know whether or not you find it satisfactory.
