### 2020-02-05 NorskRegnesentral/shapr#180 @expectopatronum
Is there a reason why you don't explicitly mention the name of the column `none`?

> The value in \code{dt[i, 1]} is the prediction value not assigned to any of the features

I actually had a hard time finding the description, because I couldn't find a mention in the README and the vignette. And I only found this paragraph by looking at the most recent commits :D
