### 2020-02-05 #2027 @expectopatronum
Everything looks fine, except maybe the documentation of `none`, I commented [here](https://github.com/NorskRegnesentral/shapr/issues/180#issuecomment-582251441)
