### 2020-02-05 NorskRegnesentral/shapr#180 @nikolase90
@expectopatronum Hehe, no it is not. Makes a lot more sense to actual use `none` instead. Will fix!
