### 2020-02-05 8c2b3c5
Fixes a small comment from review
