### 2020-02-05 #2027 @nikolase90
@expectopatronum Your last comment should be fixed with https://github.com/NorskRegnesentral/shapr/pull/191.
