### 2020-02-05 088629c
Fixes a small comment from review (#191)
