### 2020-02-05 #2027 @expectopatronum
Perfect! From my point of view the paper is good to go.
