### 2020-02-05 5b11a92
FIxed response_mod generalization
