### 2020-02-05 cb61030
starting to check dependencies on model_mat
