+++ inst/devel_scripts/true_shapley/simulation_true_shapley_withdatatable_MJ.R
@@ -33,9 +33,9 @@ parameters_list <- list(Sigma_diag = 1,
                         corr = 0,
                         mu = c(0, 0, 0),
                         beta = c(1, -1, 0, 1, 1, 1, 0.5, 0.5, 1, -1),
-                        N_shapley = 10000,
-                        N_training = 1000,
-                        N_testing = 1000,
+                        N_shapley = 1000,
+                        N_training = 100,
+                        N_testing = 100,
                         cutoff = c(-200, 0, 1, 200),
                         noise = FALSE,
                         response_mod = response_mod,
