### 2020-02-05 491fe89
added two missing dois to bib (#192)
