+++ inst/joss_paper/paper.bib
@@ -135,7 +135,8 @@ doi = {10.1214/aoms/1177728190}
   journal   = {Nature machine intelligence},
   volume    = {2},
   pages     = {56-67},
-  year      = {2020}
+  year      = {2020},
+  doi       = {10.1038/s42256-019-0138-9}
 }
 
 @article{eddelbuettel2014rcpparmadillo,
@@ -145,7 +146,8 @@ doi = {10.1214/aoms/1177728190}
   volume={71},
   pages={1054--1063},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.csda.2013.02.005}
 }
 
 @Article{ranger,
