### 2020-02-05 #2027 @martinju
@terrytangyuan This is now resolved in https://github.com/NorskRegnesentral/shapr/pull/192
