### 2020-02-05 4d9cd66
some more cleaning
continue with linear Kshap
