### 2020-02-05 99018c6
Update NEWS.md
