+++ NEWS.md
@@ -1,9 +1,9 @@
 # shapr 0.1.1
 
 * Journal of Open Source Software release
-* Improved installation instructions in README and community guidelines
+* Improved installation instructions and community guidelines in README 
 * Improved documentation
-* Some bugfixes
+* Some minor bugfixes
 
 # shapr 0.1.0
 
