### 2020-02-05 #2027 @nikolase90
@terrytangyuan New release made with https://github.com/NorskRegnesentral/shapr/pull/194. See also https://github.com/NorskRegnesentral/shapr/releases/tag/v0.1.1.
