### 2020-02-05 #2027 @kthyng
The DOI resolution is looking weird so let's wait before closing this up.
