### 2020-02-05 #2027 @kthyng
@openjournals/dev This DOI link doesn't look like it usually does while it is resolving. Any ideas? Just wait longer?
