### 2020-02-05 #2027 @arfon
This issue is being caused by `formatted_doi: XX.XXXXX/joss.XXXXX` being set in the `paper.md`. This shouldn't be in the template and a fix was applied in `rticles` in https://github.com/rstudio/rticles/pull/272

I've reprocessed this submission manually and the updated PDF should be resolving soon on the JOSS website.
