### 2020-02-05 #2027 @martinju
I would like to express my gratitude to the editor @terrytangyuan and the two reviewers @frycast  and @expectopatronum for a thorough, yet rapid, review process that led to an improved the paper, the package, and the descriptions in the README in our GitHub repo.

Thank you!
